api/requirements.txt
tensorflow>=2.16.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api>=2.16.0
matplotlib
numpy